requirements.txt
stomp.py
python-dotenv
pytz